tomat/requirements.txt
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
streamlit==1.9.2
streamlit_option_menu==0.3.2
tensorflow==2.9.1
